+++ experiments/toy_problems/experiments/consistency_regularizer/run.sh
