+++ experiments/toy_problems/experiments/consistency_with_analytical_guide_regularizer/run.sh
