+++ experiments/toy_problems/experiments/fokker_planck_regularizer/config.yaml
