+++ experiments/toy_problems/experiments/no_regularizer/run.sh
